pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "MastoBot.py"
version = "0.0.1"
description = "A generic implementation to easily create and implement bots for Mastodon"
keywords = ["mastodon", "api", "bot"]
authors = [
  { name="Erik Dreyer", email="erik.dreyer@protonmail.com" },
]
license = "GNU General Public License v3 (GPLv3)"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    'Intended Audience :: Developers',
    'Topic :: Communications',
    'Programming Language :: Python :: 3',
]
dependencies = [
  'importlib-metadata; python_version<"3.8"',
  "async-timeout==4.0.2",
  "attrs==23.1.0",
  "black==23.7.0",
  "blurhash==1.1.4",
  "boltons==23.0.0",
  "cattrs==23.1.2",
  "certifi==2023.5.7",
  "charset-normalizer==3.2.0",
  "click==8.1.6",
  "decorator==5.1.1",
  "exceptiongroup==1.1.2",
  "face==20.1.1",
  "flexicon==0.1.0",
  "greenlet==2.0.2",
  "idna==3.4",
  "Jinja2==3.1.2",
  "MarkupSafe==2.1.3",
  "Mastodon.py==1.8.1",
  "mypy-extensions==1.0.0",
  "packaging==23.1",
  "pathspec==0.11.1",
  "platformdirs==3.9.1",
  "prometheus-client==0.17.1",
  "python-dateutil==2.8.2",
  "python-magic==0.4.27",
  "regex==2023.6.3",
  "requests==2.31.0",
  "requests-cache==1.1.0",
  "six==1.16.0",
  "tomli==2.0.1",
  "typing_extensions==4.7.1",
  "url-normalize==1.4.3",
  "urllib3==2.0.4",
]

[project.urls]
"Homepage" = "https://github.com/e-dreyer/MastoBot"
"Bug Tracker" = "https://github.com/e-dreyer/MastoBot/issues"
"Pull Requests" = "https://github.com/e-dreyer/MastoBot/pulls"
"Wiki" = "https://github.com/e-dreyer/MastoBot/wiki"

[tool.setuptools.packages.find]
where = ["."]  # list of folders that contain the packages (["."] by default)
include = ["mastoBot*"]  # package names should match these glob patterns (["*"] by default)

